requirements.txt
numpy
scipy
Pillow
transformers
Flask